feat : 장바구니 기능 구현 완료(장바구니 조회,삭제 기능 추가 ,결제 완료 후 이벤트 처리 기반으로 장바구니 업데이트) chore : Product price Integer -> Long 으로 수정

+++ src/main/java/com/payflow/payflow/domain/product/Product.java
@@ -25,7 +25,7 @@ public class Product extends BaseEntity {
 
     private String name;
 
-    private Integer price;
+    private Long price;
 
     @Column(name = "file_url")
     private String fileUrl;
@@ -42,7 +42,7 @@ public class Product extends BaseEntity {
     private LocalDateTime deletedAt;
 
     @Builder
-    public Product(String name, Integer price, String fileUrl, String thumbnailUrl, String description, Long sellerId) {
+    public Product(String name, Long price, String fileUrl, String thumbnailUrl, String description, Long sellerId) {
         this.name = name;
         this.price = price;
         this.fileUrl = fileUrl;
